Build with gcc/clang -Wsign-conversion option

+++ src/libicalvcard/vcardcomponent.c
@@ -701,7 +701,7 @@ static int prop_compare(void *a, void *b)
     vcardproperty *p2 = (vcardproperty *)b;
     vcardproperty_kind k1 = vcardproperty_isa(p1);
     vcardproperty_kind k2 = vcardproperty_isa(p2);
-    int r = k1 - k2;
+    int r = (int)(k1 - k2);
 
     if (r == 0) {
         if (k1 == VCARD_X_PROPERTY) {
@@ -747,7 +747,7 @@ static int comp_compare(void *a, void *b)
     vcardcomponent *c2 = (vcardcomponent *)b;
     vcardcomponent_kind k1 = vcardcomponent_isa(c1);
     vcardcomponent_kind k2 = vcardcomponent_isa(c2);
-    int r = k1 - k2;
+    int r = (int)(k1 - k2);
 
     if (r == 0) {
         int i;
